+++ CHANGELOG.md
@@ -1,7 +1,30 @@
+## 0.3.1 (2023-03-29)
+
+### Fix
+
+- req.txt
+- import error
+
+## 0.3.0 (2023-03-29)
+
+### Feat
+
+- tasks
+- nearest
+- end /reg command refactor
+- changed storage from redis to postgresql
+
+### Fix
+
+- moree fixes for handlers birthdays
+- added default timezone
+- 06 -> 00:00
+
 ## 0.2.0 (2022-10-20)
 
 ### Feat
 
+- start adding BIG FEAT for anon congratulations
 - addeing sex
 - new test handler + some fixes for registr + adding dependency injector for cache
 - alpha version of bot